@@ -1,12 +1,12 @@
 package nl._42.heph;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import nl._42.heph.builder.FindEntityMethodNotImplementedFixtures;
 import nl._42.heph.shared.AbstractSpringTest;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class FindEntityMethodNotImplementedTest extends AbstractSpringTest {
