@@ -1,11 +1,11 @@
 package nl._42.heph.builder;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import nl._42.heph.domain.AncientTribe;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class AncientTribeFixturesTest {
 
