@@ -1,7 +1,7 @@
 package nl._42.heph.builder;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotSame;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotSame;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -12,7 +12,7 @@ import nl._42.heph.domain.OrganizationRepository;
 import nl._42.heph.domain.Person;
 import nl._42.heph.shared.AbstractSpringTest;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class PersonFixturesTest extends AbstractSpringTest {
