@@ -1,22 +1,19 @@
 package nl._42.heph.shared;
 
 import nl._42.database.truncator.DatabaseTruncator;
-import nl._42.heph.ApplicationConfig;
+import nl._42.heph.Application;
 
-import org.junit.Before;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.boot.test.context.SpringBootTest;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = { ApplicationConfig.class })
+@SpringBootTest(classes = Application.class)
 public abstract class AbstractSpringTest {
 
     @Autowired
     private DatabaseTruncator databaseTruncator;
 
-    @Before
+    @BeforeEach
     public void cleanUp() throws Exception {
         databaseTruncator.truncate();
     }
