@@ -4,6 +4,7 @@ import nl._42.heph.AbstractBuilder;
 import nl._42.heph.domain.Person;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -12,9 +13,10 @@ public class PersonFixtures extends AbstractBuilder<Person, PersonBuildCommand> 
     public static final String EXPECTED_NAME = "Sjaak";
 
     @Autowired
+    @Lazy
     private OrganizationFixtures organizationFixtures = new OrganizationFixtures();
-
     @Autowired
+    @Lazy
     private WorkspaceFixtures workspaceFixtures = new WorkspaceFixtures();
 
     @Override
